Re-implemented 'compose' function.
This is closer to the definition of function composition and much closer to what
I had intended to write in the first place: we compose functions and return a
function! The rub here is this implementation can only handle arity-one in the
lead function of the composition.

This partially addresses #20.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>