tmap natively uses EEP0018 Object form C.
tmap is meant to work over common erlang trees. That, in the author's
experience, means json and gb_tree. While the later is not supported yet, the
former was supported but poorly. EEP0018--the proposal for json nifs--sets out
three object formats; the Cth is:

    C. {[{K1,V1},...,{Kn,Vn}]} Like B, but there isn't any need for a tag.

The EEP goes on to say of this form: "It has recently had supporters." In the
time since that writing it has become the norm, being the native format of
[jiffy](https://github.com/davisp/jiffy) and a native format of
[mochijson2](https://github.com/mochi/mochiweb/blob/master/src/mochijson2.erl).

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>